Add API setting to require capture ID to resolve alias
This introduces a new API setting called ``require_captid_to_resolve``
that controls whether to require a capture ID match to find an alias.
This skips the fallback lookup that's done without capture ID, in order
to reduce the number of cases where a service in a network that
overlaps with an unrelated alias, receives that alias as a label.

Enabling this new setting effectively enables ``add_captid_to_resolve``
as well.